Refactored views structure to handle versions of api
