+++ kpi/views/current_user.py
