+++ kpi/views/no_update_model.py
