+++ kpi/views/token.py
@@ -9,7 +9,6 @@ from rest_framework.authtoken.models import Token
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-
 class TokenView(APIView):
     def _which_user(self, request):
         """
