+++ kpi/views/v1/__init__.py
@@ -1,2 +1,19 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
+
+from .asset import AssetViewSet
+from .asset_file import AssetFileViewSet
+from .asset_snapshot import AssetSnapshotViewSet
+from .asset_version import AssetVersionViewSet
+from .authorized_application_user import AuthorizedApplicationUserViewSet
+from .collection import CollectionViewSet
+from .export_task import ExportTaskViewSet
+from .hook_signal import HookSignalViewSet
+from .import_task import ImportTaskViewSet
+from .object_permission import ObjectPermissionViewSet
+from .one_time_authentication_key import OneTimeAuthenticationKeyViewSet
+from .sitewide_message import SitewideMessageViewSet
+from .submission import SubmissionViewSet
+from .tag import TagViewSet
+from .user import UserViewSet
+from .user_collection_subscription import UserCollectionSubscriptionViewSet
