+++ kpi/views/v1/object_permission.py
@@ -8,7 +8,7 @@ from kpi.constants import PERM_SHARE_SUBMISSIONS
 from kpi.filters import KpiAssignedObjectPermissionsFilter
 from kpi.models import ObjectPermission
 from kpi.serializers import ObjectPermissionSerializer
-from .no_update_model import NoUpdateModelViewSet
+from kpi.views.no_update_model import NoUpdateModelViewSet
 
 
 class ObjectPermissionViewSet(NoUpdateModelViewSet):
